Shell out app and wear modules into android-application, app-impl, app-test-fixtures, and wear-impl
Fixes #2955
