+++ android-application/build.gradle.kts
@@ -1,0 +1,52 @@
+/*
+ * Copyright 2025 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import com.alexvanyo.composelife.buildlogic.FormFactor
+import com.alexvanyo.composelife.buildlogic.configureGradleManagedDevices
+import com.android.build.api.dsl.KotlinMultiplatformAndroidDeviceTestCompilation
+
+plugins {
+    alias(libs.plugins.convention.kotlinMultiplatform)
+    alias(libs.plugins.convention.androidLibrary)
+    alias(libs.plugins.convention.androidLibraryCompose)
+    alias(libs.plugins.convention.detekt)
+    alias(libs.plugins.gradleDependenciesSorter)
+    alias(libs.plugins.metro)
+}
+
+kotlin {
+    androidLibrary {
+        namespace = "com.alexvanyo.composelife.androidapplication"
+        minSdk = 23
+        compilations.withType(KotlinMultiplatformAndroidDeviceTestCompilation::class.java) {
+            instrumentationRunner = "com.alexvanyo.composelife.test.InjectTestRunner"
+        }
+        configureGradleManagedDevices(setOf(FormFactor.Mobile), this)
+    }
+
+    sourceSets {
+        val commonMain by getting {
+            dependencies {
+                implementation(libs.androidx.compose.runtime)
+                implementation(libs.androidx.lifecycle.runtime)
+                implementation(projects.injectScopes)
+                implementation(projects.processLifecycle)
+                implementation(projects.strictMode)
+                implementation(projects.updatable)
+            }
+        }
+    }
+}
