+++ app-impl/src/androidMain/AndroidManifest.xml
@@ -1,0 +1,56 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+     Copyright 2022 The Android Open Source Project
+
+     Licensed under the Apache License, Version 2.0 (the "License");
+     you may not use this file except in compliance with the License.
+     You may obtain a copy of the License at
+
+          http://www.apache.org/licenses/LICENSE-2.0
+
+     Unless required by applicable law or agreed to in writing, software
+     distributed under the License is distributed on an "AS IS" BASIS,
+     WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+     See the License for the specific language governing permissions and
+     limitations under the License.
+-->
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+
+    <uses-sdk tools:overrideLibrary="androidx.xr.arcore,androidx.xr.compose,androidx.xr.runtime,androidx.xr.runtime.openxr,androidx.xr.scenecore,com.google.ar.imp.apibindings" />
+
+    <application
+        android:name="com.alexvanyo.composelife.ComposeLifeApplication"
+        android:allowBackup="true"
+        android:enableOnBackInvokedCallback="true"
+        android:fullBackupOnly="true"
+        android:icon="@mipmap/ic_launcher"
+        android:label="@string/app_name"
+        android:roundIcon="@mipmap/ic_launcher_round"
+        android:supportsRtl="true"
+        android:theme="@style/Theme.ComposeLife.Starting"
+        tools:targetApi="34">
+
+        <profileable
+            android:shell="true"
+            tools:targetApi="29" />
+
+        <activity-alias
+            android:name="Launcher"
+            android:exported="true"
+            android:targetActivity="com.alexvanyo.composelife.MainActivity">
+            <intent-filter>
+                <action android:name="android.intent.action.MAIN" />
+
+                <category android:name="android.intent.category.LAUNCHER" />
+                <category android:name="android.intent.category.DEFAULT" />
+            </intent-filter>
+        </activity-alias>
+
+        <activity
+            android:name="com.alexvanyo.composelife.MainActivity"
+            android:configChanges="allKnown"
+            android:exported="false" />
+    </application>
+
+</manifest>
