+++ app-impl/src/androidMain/res/drawable/ic_splash.xml
