+++ app-test-fixtures/README.md
@@ -1,0 +1,4 @@
+# app
+
+An application module defining the mobile app, which displays a configurable, interactable Game of
+Life simulation.
