+++ settings.gradle.kts
@@ -64,10 +64,13 @@ include(":algorithm")
 include(":algorithm-benchmark")
 include(":algorithm-molecule-tests")
 include(":algorithm-test-resources")
+include(":android-application")
 include(":app")
 include(":app-baseline-profile-generator")
 include(":app-benchmark")
 include(":app-compat-sync")
+include(":app-impl")
+include(":app-test-fixtures")
 include(":clock")
 include(":data")
 include(":data-test-resources")
@@ -129,6 +132,7 @@ include(":updatable")
 include(":web-app")
 include(":wear")
 include(":wear-baseline-profile-generator")
+include(":wear-impl")
 include(":wear-watchface")
 include(":wear-watchface-wff-resources")
 include(":wear-watchface-wff-version-1")
