+++ wear-impl/src/androidMain/kotlin/com/alexvanyo/composelife/wear/ComposeLifeApplication.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright 2022 The Android Open Source Project
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.alexvanyo.composelife.wear
+
+import com.alexvanyo.composelife.BaseComposeLifeApplication
+import com.alexvanyo.composelife.scopes.GlobalScope
+import dev.zacsweers.metro.DependencyGraph
+import dev.zacsweers.metro.createGraph
+
+class ComposeLifeApplication : BaseComposeLifeApplication<GlobalGraph>(::createGraph)
+
+@DependencyGraph(GlobalScope::class)
+interface GlobalGraph
